Remove redundant files during up-to-date verification
